requirements.txt
# Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
flask==3.0.0

# AI和Agent
langgraph==0.0.26
langchain-anthropic==0.1.8
anthropic==0.18.0
langchain-core==0.1.40

# 微信SDK
wechatpy==1.8.18

# 数据库
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.1

# 任务队列
celery==5.3.4
flower==2.0.1
kombu==5.3.5

# 工具
pyyaml==6.0.1
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0

# 监控和日志
prometheus-client==0.19.0
sentry-sdk==1.40.0

# 其他
python-multipart==0.0.6
aiofiles==23.2.1
jinja2==3.1.3
